+++ cmd/pkg/update_test.go
@@ -10,89 +10,74 @@ import (
 	"testing"
 )
 
-type UpdateTestData struct {
-	TestData
-
-	releases map[string]string
-}
-
 func TestUpdateCommand(t *testing.T) {
-	tests := []UpdateTestData{
+	tests := []TestData{
 		{
-			TestData: TestData{
-				name:            "Should bump the Ref field for the specified packages",
-				args:            []string{"app-hello", "load-balancing-alb-main"},
-				testdataRootDir: "testdata/update/bump-ref-field",
-				expectError:     false,
-				expectedFiles: []string{
-					"packages.yml",
-					"config/app-hello.yml",
-					"config/common-config.yml",
-				},
-			},
+			name:            "Should bump the Ref field for the specified packages",
+			args:            []string{"app-hello", "load-balancing-alb-main"},
+			testdataRootDir: "testdata/update/bump-ref-field",
 			releases: map[string]string{
 				"app":                "v9.0.0",
 				"load-balancing-alb": "v4.0.0",
 				"app-common":         "v7.0.0",
 			},
+			expectError: false,
+			expectFiles: []string{
+				"packages.yml",
+				"config/app-hello.yml",
+				"config/common-config.yml",
+			},
 		},
 		{
-			TestData: TestData{
-				name:            "Should bump the Ref field only for semver-version package Refs",
-				args:            []string{},
-				testdataRootDir: "testdata/update/bump-ref-field-semver-only",
-				expectError:     false,
-				expectedFiles: []string{
-					"packages.yml",
-				}},
+			name:            "Should bump the Ref field only for semver-version package Refs",
+			args:            []string{},
+			testdataRootDir: "testdata/update/bump-ref-field-semver-only",
+			expectError:     false,
+			expectFiles: []string{
+				"packages.yml",
+			},
 			releases: map[string]string{
 				"app": "v9.0.0",
 			},
 		},
 		{
-			TestData: TestData{
-				name:            "Should bump schema version in var files",
-				args:            []string{"app-hello"},
-				testdataRootDir: "testdata/update/bump-schema-version",
-				expectError:     false,
-				expectedFiles: []string{
-					"packages.yml",
-					"config/app-hello.yml",
-					"common-config.yml",
-				},
+			name:            "Should bump schema version in var files",
+			args:            []string{"app-hello"},
+			testdataRootDir: "testdata/update/bump-schema-version",
+			expectError:     false,
+			expectFiles: []string{
+				"packages.yml",
+				"config/app-hello.yml",
+				"common-config.yml",
 			},
 			releases: map[string]string{
 				"app": "v9.0.1",
 			},
 		},
 		{
-			TestData: TestData{
-				name:            "Should migrate schema declaration from dir based to HTTPS based",
-				args:            []string{"app-hello"},
-				testdataRootDir: "testdata/update/migrate-schema-declaration-format",
-				expectError:     false,
-				expectedFiles: []string{
-					"packages.yml",
-					"config/app-hello.yml",
-					"common-config.yml",
-				},
+			name:            "Should migrate schema declaration from dir based to HTTPS based",
+			args:            []string{"app-hello"},
+			testdataRootDir: "testdata/update/migrate-schema-declaration-format",
+			expectError:     false,
+			expectFiles: []string{
+				"packages.yml",
+				"config/app-hello.yml",
+				"common-config.yml",
 			},
 			releases: map[string]string{
 				"app": "v9.0.1",
 			},
 		},
 		{
-			TestData: TestData{
-				name:            "Should update ok packages recursively",
-				args:            []string{"--recursive"},
-				testdataRootDir: "testdata/update/recursive",
-				expectError:     false,
-				expectedFiles: []string{
-					"app-common/packages.yml",
-					"app-common/config.yml",
-					"app-hello/packages.yml",
-					"app-hello/config.yml",
-				},
+			name:            "Should update ok packages recursively",
+			args:            []string{"--recursive"},
+			testdataRootDir: "testdata/update/recursive",
+			expectError:     false,
+			expectFiles: []string{
+				"app-common/packages.yml",
+				"app-common/config.yml",
+				"app-hello/packages.yml",
+				"app-hello/config.yml",
 			},
 			releases: map[string]string{
 				"app":        "v9.0.0",
@@ -104,7 +89,7 @@ func TestUpdateCommand(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			// Given
-			testDir, err := os.Getwd()
+			testWorkingDirectory, err := os.Getwd()
 			require.NoError(t, err)
 
 			ghReleases := &GitHubReleasesMock{
@@ -116,19 +101,24 @@ func TestUpdateCommand(t *testing.T) {
 			tempDir, err := os.MkdirTemp(os.TempDir(), "ok-"+tt.name)
 
 			// Remove temp dir after test run
-			defer func(path string) {
-				err := os.RemoveAll(path)
-				require.NoError(t, err)
-			}(tempDir)
+			if !tt.keepTempDir {
+				defer func(path string) {
+					err := os.RemoveAll(path)
+					require.NoError(t, err)
+				}(tempDir)
+			}
 
 			require.NoError(t, err)
 
 			fmt.Println("tempDir: ", tempDir)
-			copyTestdataRootDirToTempDir(t, tt.TestData, tempDir)
+			copyTestdataRootDirToTempDir(t, tt, testWorkingDirectory, tempDir)
 			command.SetArgs(tt.args)
 
 			err = os.Chdir(tempDir) // Works, but disables the possibility for parallel tests.
 			require.NoError(t, err)
+			defer func() {
+				err = os.Chdir(testWorkingDirectory)
+			}()
 
 			// When
 			err = command.Execute()
@@ -140,16 +130,17 @@ func TestUpdateCommand(t *testing.T) {
 			}
 			require.NoError(t, err)
 
-			err = os.Chdir(testDir)
+			err = os.Chdir(testWorkingDirectory)
 			require.NoError(t, err)
 
 			// Compare package manifest file
-			for _, file := range tt.expectedFiles {
-				actualBytes, err := os.ReadFile(filepath.Join(tempDir, file))
+			for _, expectedFile := range tt.expectFiles {
+				actualBytes, err := os.ReadFile(filepath.Join(tempDir, expectedFile))
 				require.NoError(t, err)
 				actual := string(actualBytes)
 
-				expectedBytes, err := os.ReadFile(filepath.Join(tt.testdataRootDir, "expected", file))
+				expectedFileFullPath := filepath.Join(tt.testdataRootDir, "expected", expectedFile)
+				expectedBytes, err := os.ReadFile(expectedFileFullPath)
 				require.NoError(t, err)
 				expected := string(expectedBytes)
 
@@ -157,12 +148,4 @@ func TestUpdateCommand(t *testing.T) {
 			}
 		})
 	}
-}
-
-type GitHubReleasesMock struct {
-	LatestReleases map[string]string
-}
-
-func (g *GitHubReleasesMock) GetLatestReleases() (map[string]string, error) {
-	return g.LatestReleases, nil
 }
